clean.sh aware of docker-machine

+++ clean.sh
@@ -1,8 +1,12 @@
 #!/usr/bin/env bash
+source lib.sh
 
-docker rm -f $(docker ps -aq)
+#docker rm -f $(docker ps -aq)
+set - x
+[ -z "$DOCKER_HOST" ] && sudo rm -rf crypto-config/ || docker-machine ssh ${DOCKER_MACHINE_NAME} sudo rm -rf crypto-config
+set +x
+[ "${DOCKER_MACHINE_NAME}" == "orderer" ]  && EXECUTE_BY_ORDERER=1 runCLIWithComposerOverrides down || runCLIWithComposerOverrides down
 docker volume prune -f
-sudo rm -rf crypto-config/
 docker rmi -f $(docker images -q -f "reference=dev-*")
 #docker rmi -f $(docker images -q -f "reference=olegabu/fabric-starter-client")
 docker network rm `(docker network ls -q)`
